backend/requirements.txt
fastapi
openai
uvicorn
python-dotenv
pytest
pytest-asyncio
httpx
pytest-mock